ユーザーページ以外のページ遷移まで確認
